Enable Python 2.6 compatibility.
In case the Python version is < 2.7 this patch enables the use

backport_collections
unittest2

Both are available on pip.

On Python 2.6 mmap throws a different type of error. With this patch this error is correctly caught.

With these changes the test suite passes without errors (some tests are skipped but this is also true on my Python 2.7 system).