Fix `chess.svg` module
This patch is the second (and hopefully the last) attempt to correctly
document some arguments of the `chess.svg.board` function. Also, the
patch fixes the value of the `DEFAULT_STYLE` constant so that the string
is on one line only and is not triple quoted and that the ugly backslash
can be removed. There's no point for a constant's value to be laid out
so that something can be added to it. Remember, it's a constant.