Update pgn.py #2
This little patch changes capital initials to normal initials for the text "notation annotation glyphs", because everywhere else in the code (in other places) we don't say, like, Standard Algebraic Notation (SAN), we say it without all the initials in each word capitalized, which is right. Here is the same thing.