Fixed draw handling and added `resign` for engine
The main issue was that `move_received` had to be set in order to
continue with the `go()` function. So some dummy responses had to be
added and `move_receieved` had to be set when the engine replied with
`resign` or with `offer draw` and we did decide to draw.