Update xboard.py #2
A tip from mister Dugović led me to revisit my patch and I fixed some things for which I wasn't 100% sure if I should fix and how should I fix them. I've done more research and now some sentences are correctly written. Also, a draw is not expired but declined, so I fixed that. A big patch and a lot of time, but it will benefit the documentation immensely when it goes online (when the xboard module is ready for prime time).