Update setup.py
Hey! Thanks for letting us use your great work. I would suggest you change the name of the module in setup.py to "pythonchess". This way we would be able to pip install it and import the module without being in the main directory. Import statements with a "-" won't work. Thanks, and keep the good work up!
Merlin