Fix SQUARE_SIZE constant to 50
This constant is currently wrong. The default size of the SVG chessboard is 400, divide that with 8 (as we have 8 files/ranks) and the result is 50, so the square size is 50. I have noticed in my app that it is off for exactly 5 pixels. Now I finally found the culprit. Please compile a new version of python-chess.