Rename variable 't' to something more descriptive
I've renamed the variable 't', which is very undescriptive. Every variable should be properly named so that its name alone describes why is it there and what data does it hold. I hope I understand the code enough to give the variable 't' a proper name. If not, please rename it yourself.