Simplify release.py module
Call the int() function just once. Current code calls the int() function twice in the case  master is behind.

Furthermore, the shebang line is improved, visually and functionally. Calling ‘env’ is preffered. It ensures that the Python interpreter is found (by the Linux ‘env’ command) even if installed on some custom path. This shebang line needs to change in every module that exists for the python-chess project. I can make a PR for this.