Remove the obsolete first line
Now that python-chess supports only Python 3.6+ (for some time
now), we don’t need this first line anymore. Python 3 does not need
this line because UTF-8 is universal. In any other module in the python-chess project it
is obsolete. I am removing it in this module. Other modules should follow suit.