Simplify unittest.skipIf() decorator condition
This change provides the same result, but it is more readable. Also, someone on StackOverflow ensured that checking for the OS version is more reliable by using platform.system(): https://stackoverflow.com/questions/1854/what-os-am-i-running-on/58071295#58071295

Furthermore, the current Python version comparison is the same now and not complicated. It's simpler to compare the Python version as being less than version 3.8.1 as opposed to less than or equal to 3.8.0 and less than 3.8.1.