Add a GET_INDEX dictionary
It is intended to be used like this:
chess.Move(GET_INDEX["e2"], GET_INDEX["e4"])

In some situations, it may be convenient to have a GET_INDEX dictionary if the programmer is getting square data in string form (str type) and wants to pass that directly into chess.Move(). And since chess.Move() accepts only indexes (int type), this could be a convenient feature.

I couldn't think of a better name for this. Maybe GET_INDEX is not so good as a name. Maybe just INDEX? It's shorter. I don't know.