Fix push_san() docstring
Fixed a typo where it says "on the the", clearly an overlook from 5 years ago. The correct wording is "onto the", talking about a move stack. As stated on GrammarBook.com: In general, use "onto" as one word to mean “on top of,” “to a position on,” “upon.” And since we push moves "on top of" the stack, I fixed the wording accordingly.