Fix some peculiarities in the __init__
Fixed some peculiarities I found in the __init__. As I noticed throughout the code, python-chess does not use a comma before "and" or "or", so I fixed those occurences and have tweaked one of the sentences a bit to be more understandable.

Also, the STARTING_FEN and STARTING_BOARD_FEN constants have a triple-quoted string on a line below them. That's kind of strange. Looks like a docstring, but it actually isn't. I converted that into a regular comment for each constant, also tweaked the sentences a bit.

Then another thing I did was fixed the docstring of the Board() subclass.

@niklasf, I hope you like my ideas.