document `legal_drop_squares` and remove unnecessary copies in `_CrazyhouseBoardState`
`legal_drop_squares` was not documented yet not used in the code source and considering it returns a `chess.SquareSet` I think it was intented to be exposed to the users.

when restoring a previous position form a `_CrazyhouseBoardState` it's not necessary to make a copy of the pocket since the instance is discarded. Also it's not intented to be publicly used.

All test passed.