Adjust time used for pondering time
The time between a move being received (whether by position or
ponderhit) and bestmove being sent back is subtracted from the time in
the initial Limit paramter in the play() call. If there is a clock
increment, it is added to the ponder time. If there is a remaining_moves
value in the limit, it is decremented.

All of these operations should give the engine a good idea of the
current state of the clock after picking a move. The only deficiency is
that if remaining_moves becomes zero, then the ponder time will look
like the time for the rest of the game. Although, since game periods are
commonly 40 moves, this is still a decent approximation.