Add ascii board as description of svg
The tag <desc> in SVG can store a textual description, eg. for better
accessibility. For the chess board there could be the string
representation (basically str(chess.Board()) and enclosed in <pre> it is
readable in text browsers like links or w3m when the svg is embedded in
html.