Add optional identifier to Limit instances
This new field in the Limit class is used to signal when the time
control for a game has changed. XBoard engines expect to
receive a `level` or `st` command to indicate the time control
at the start of the game, and then only `time` and `otim`
commands before each subsequent move to update the engine's
internal clocks. The `level` and `st` commands should only be
resent when the time control changes--like in FIDE tournament
time controls. During a single time control, the same `clock_id`
should be used for every Limit instance. This field has a similar
function to the `game` parameter of the `play()` method.