Create new function for parsing engine output
The function next_token() extracts the next word from a string and
removes the leading and trailing whitespace around that word. It
also returns the remainder of the string unchanged. Using this function
instead of split() gets rid of extraneous whitespace when searching
for keywords and preserves whitespace in strings generated by
users and their engines.